Cleanup and full ci
