+++ .github/workflows/zip.yml
@@ -46,7 +46,7 @@ jobs:
 
   quickstart_framework_abtesting:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+#    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -91,7 +91,7 @@ jobs:
 
   quickstart_framework_auth:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -135,7 +135,7 @@ jobs:
 
   quickstart_framework_config:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -176,7 +176,7 @@ jobs:
 
   quickstart_framework_crashlytics:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -226,7 +226,7 @@ jobs:
 
   quickstart_framework_database:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -271,7 +271,7 @@ jobs:
 
   quickstart_framework_dynamiclinks:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -321,7 +321,7 @@ jobs:
 
   quickstart_framework_firestore:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -364,7 +364,7 @@ jobs:
 
   quickstart_framework_inappmessaging:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -410,7 +410,7 @@ jobs:
 
   quickstart_framework_messaging:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
@@ -455,7 +455,7 @@ jobs:
 
   quickstart_framework_storage:
     # Don't run on private repo.
-    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
+##    if: github.event_name == 'schedule' && github.repository == 'Firebase/firebase-ios-sdk'
     needs: package
     env:
       plist_secret: ${{ secrets.GHASecretsGPGPassphrase1 }}
